+++ customer_portal/src/epilot/models/components/actionwidget.py
@@ -32,6 +32,7 @@ class ActionWidgetType(str, Enum):
     ENTITY_WIDGET = 'ENTITY_WIDGET'
     TEASER_WIDGET = 'TEASER_WIDGET'
     DOCUMENT_WIDGET = 'DOCUMENT_WIDGET'
+    PAYMENT_WIDGET = 'PAYMENT_WIDGET'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)